bsp/irq-server: Fix install/remove
Do not wait for the interrupt server while owning the allocator lock.
This could lead to deadlock in case one of interrupt service routines or
user extensions want to obtain the allocator mutex as well.  Instead let
the interrupt server do the install/remove job entirely on behalf of the
requesting task.