score: Inline some SMP lock operations by default
The SMP ticket lock release turned out to be suitable for inlining, e.g.
a hand full of instructions, no branches.

The changes in the screen files do not reflect the changes due to this
commit.  However, they are now up to date.  Obtained on a T4240 running
at 1.5GHz using GCC 7.0.0 20161108..