score: Fix user extensions order
Use forward and reverse order for initial and dynamic extensions.  This
is the behaviour documented in the C Users Guide.  Change thread
terminate order to backward to be in line with the thread delete order.
Change fatal error order to forward to ensure that initial extensions
are called first due the peculiar execution context of fatal error
extensions, see _Terminate() documentation.

Update #2692.