score: Improve scheduler helping protocol
Only register ask for help requests in the scheduler unblock and yield
operations.  The actual ask for help operation is carried out during
_Thread_Do_dispatch() on a processor related to the thread.  This yields
a better separation of scheduler instances.  A thread of one scheduler
instance should not be forced to carry out too much work for threads on
other scheduler instances.

Update #2556.