bsps/arm: Remove DWT based clock.
It seems that the DWT CYCCNT does not advance when the CPU waits on a
WFI instruction.

That leads to the effect that for example on the atsamv BSP a sleep(1)
needs something in the range of a few minutes (depending on the
configured systick). A debugger might disables some deep sleep modes so
that the problem only appears if the application is executed without a
debugger.