bsps: Remove AC_CONFIG_SRCDIR()
This AC_CONFIG_SRCDIR() is just a sanity check in this insane build
system.  Since all content of
c/src/lib/libbsp/@RTEMS_CPU@/@RTEMS_BSP_FAMILY@ is bound to be moved it
makes no sense to keep it.

This patch is a part of the BSP source reorganization.

Update #3285.