Remove INTERNAL_ERROR_INTERRUPT_STACK_TOO_SMALL
The configured interrupt stack size (CONFIGURE_INTERRUPT_STACK_SIZE) is
checked against the minimum task stack size.  The minium tasks task
stack size is also a configuration option
(CONFIGURE_MINIMUM_TASK_STACK_SIZE).  So, this check does not really
help in case of configuration errors.  In addition, the interrupt stack
is also re-used as the initialization stack in most BSPs.  It is
probably better to use a stack checker to detect problems.

Update #3459.