Add rtems_interrupt_server_request_set_vector()
By default, the interrupt vector of an interrupt server request is set
to a special value which is outside the range of vectors supported by
the interrupt controller hardware.

Add rtems_interrupt_server_request_set_vector() to set the interrupt
vector in an interrupt server request.

Calls to rtems_interrupt_server_request_submit() will disable the interrupt
vector of the request.  After processing of the request by the interrupt
server the interrupt vector will be enabled again.