telnetd: Remove keep stdio feature
The Telnet service started via rtems_telnetd_start() had a keep stdio
feature.  This just created a task and executed the command function in
a loop.  For this kind of service we do not library support.  This can
be done by an application task on its own.  Remove this feature and
provide only the real Telnet server functionality.

Use syslog() for error and status messages.

Add test program for the Telnet server.

Close #3542.