spconfig02: Check object methods in default config
Ensure that the creation of Classic API objects fails with the expected
status code in the default configuration.

Ensure that the deletion of Classic API objects fails with the expected
status code in the default configuration if the identifier is invalid.

Ensure that only the expected objects are present in the default
configuration via rtems_object_get_classic_name().