psxtimes01: Fix sporadic failures
It is not enough to just wait for at least the first clock tick to get
non-zero times.  In _times() we round down, so the uptime must be
greater than or equal to the clock tick interval.  We do not know when
the timecounter is started during the system initialization.  We have to
wait until at least two ticks happened.